findpkg: query primitives for *packages.Package
